Add filter to cli command 'proposals'
Currently when using the cli and running command `proposals` we get all
the proposals back.  It would be nice to be able to narrow down the
search i.e using a filter.  We can filter based on the country and also
the provider ID.

Add filter term to command `proposals`.  No documentation added.  Needs
extending to include something in the output of `help` to show that this
feature exists.

Signed-off-by: Tobin C. Harding <me@tobin.cc>