Conditionally add broker port number
Currently we add the hard coded broker port number 4222
unconditionally.  If a user wishes to configure nats to use a different
port number then we need to accept this port number as part of the
configuration for running myst.  If we conditionally add the port number
then the following is supported

     bin/run --broker-addres="localhost:3333"

It seems that nats library should handle an address string that does not
contain the port number by using the default port but our e2e tests fail
if the port number is not present.

Conditionally add broker port number to the address string.

Signed-off-by: tcharding <me@tobin.cc>