cli: Warn if unknown sub-command passed to `identities`
Currently if a user passes an unknown sub-command to the `identities`
command we fail silently.  We check for zero args and provide usage, we
might as well check for incorrect args and provide usage also.  This
gives us `identities help` for free.

Add check on the sub command for `identities` and print usage string if
incorrect.

Signed-off-by: tcharding <me@tobin.cc>