bootstrap now checks for presence of node before starting
We're running into an issue where the node is being started multiple times. This causes various issues, namely boltdb lock races. To fix it, introducing an early attempt to bind to the tequilapi port. If the bind suceeds, we'll now that a node isn't running. Well, at least on the same port.