Ghost sessions will now disappear in due time
We were using a noop balance tracker in openvpn case which resulted in ghost sessions hanging on the provider side if the consumer failed to send a session destroy request. These will now dissappear in due time(2 hours). The noop payment provider should be safe to remove as most of the consumer will have probably upgraded by now to a version which supports payments.