Openvpn server now reports session data transfer statistics
The openvpn service will now publish events on statistics data change. These will trigger an event in the session storage, which will in turn update the required session. It will notify all the subscribers of a session change after that. This will in turn update the state and push the change through the events endpoint via sse.