Implemented off chain payment message dialog between consumer and provider
This commit introduces the first iteration of the new payment version for node.

The provider issues invoices on a regular basis. These invoices are sent to the consumer,
who in turn issues a message to the provider with a signed promise. The provider then takes the
promise and asks the accountant to provide an accountant promise to the provider.