fixes a race condition where ssdp byes are being sent multiple times with the ssdp client already being closed
The issue is caused by multiple stop calls, hence the second call tries to send a bye when the ssdp client is already closed.

Moving the relevant code to the once func will solve the issue. However, there needs to be a better look at our stopping routines.
We need to figure out why the stops are being called multiple times. This is out of scope for this commit though.

Fixes #1333