remove redundant R recovery
with the agreement ID's being random, the current R recovery logic does not do what it's supposed to do.

With the changes in invoice tracking the recovery should not be required, therefore removing it permanently.

A mechanism similar to this one will be needed once accountant starts returning consistent errors that we can handle.