Standardize datasize units
To avoid ambiguity, name the units per specification (we use powers of 2) as opposed to common usage (deprecated).

https://en.wikipedia.org/wiki/Kilobyte#Base_2_(1024_bytes)

In December 1998, the IEC addressed such multiple usages and definitions by creating prefixes such as kibi, mebi, gibi, etc., to unambiguously denote powers of 1024.[10] Thus the kibibyte, symbol KiB, represents 210 = 1024 bytes. These prefixes are now part of the International System of Quantities. The IEC further specified that the kilobyte should only be used to refer to 1000 bytes.