Include a info subcommand (#2865)
Added a command which shows your current connection information.
Current output if everything goes right is this:
```
[INFO] IP: 78.157.74.92
[INFO] Status: Connected
[INFO] Location: Vilnius, LT (residential - Penki)
[INFO] Using Identity: 0xe158f8bd585619a6f2f74527df6518074acde542
[INFO] Session ID: 9ac86d32-9e69-4a0d-bef1-08d57462b294
[INFO] Proposal: Id: 1 , Provider: 0x7a856876bb6fcc8765261e9a7e29463824bd6742, Country: CA
[INFO] Duration: 21m0s
[INFO] Transfered: 0b/0b
[INFO] Throughput: 0bs/0bs
[INFO] Spent: 0.000200MYST
```

My reasoning for the complicated data setting and printing: We want to shows a similar output to the user every singe time he runs the command, even if something fails. So now if something fails we just replace it with an `error`:
```
[WARNING] IP: No data retrieved
[INFO] Status: Connected
[INFO] Location: Vilnius, LT (residential - Penki)
[WARNING] Using Identity: No data retrieved
.....
```

This was the easiest way I found to do it with the least amount of code repetition.