Attempt to avoid ghost sessions in NodeUI
We have been getting some bug reports that SSE state that we push contain long-running sessions with 0 traffic. It is possible that Session `Created` and `Removed` could be fired almost the same time but due to `SubscribeAsync` will arrive in reverse order that permanently adding a session to pushed SSE state which in turn will display a ghost session in NodeUI that is not sending/receiving traffic.

Signed-off-by: Mantas Domaševičius <mantas@mysterium.network>