Change registration workflow to improve offchain flow
* Registration requests must be sent after sending offchain addition.

Because of how offchain works, `GetRegistrationStatus` can not only get the registration
status but also register an identity. This forces us to adapt to this workflow and return status 200 - registered
if user becomes registered as offchain.

* Check InProgress registrations on hermes also.

If we send a registration request and hermes is down for some reason but the target is offchain
we will get in trouble. We will try to register onchain, marking the registration as in progress
and we will not check hermes again for hours, potentialy bricking the new client.