Handle missing consumer promise when setting withdrawal amount
If latest promise is invalid, issue a new promise with amount equal to settled+amount to withdraw. This way we can still recover and withdraw even though provider is has corrupted data.
On hermes side we will also be able to recover withdraw promises using the help of this change.