Handle registration status error for state subscribers (#4414)
If there are connection issues on initial load, apps will assume that user is unregistered forcing them to the topup/onboarding screen. Now they can rely on state Unknown to be sure that something failed and should be retried.

This change will need to handled in both android and desktop teams when they upgrade node.