Retry withdrawal if it's still valid and possible
Rely on transactor response codes being correct and retry
a withdrawal if it failed but the request itself was valid.

While retrying a few times for 5 minutes, doesn't completely solve our
issue, this and the previous withdrawal PR should remedy 90% or more of
the withdrawal problems being reported.

Solving this completely is rather difficult because we would have to
somehow track this promise and make sure we're resending it, but as it's
mentioned in the comment block in code, sending a settlement after 5
minutes will produce different results then shown in the UI as fees will
change and sometimes they will change dramatically.

Small issue which was not resolved yet: Every other withdrawal might
fail with no money being transferred. That is because we've stopped
lying to the UI and actually report failures now.