Ignore withdrawal promises on initial load
This fixes node earnings resetting to 0 on start up.
What would happen is on startup we will pull 2 promises, one of them
will be a regular promise and another a withdrawal. If we first pulled a
regular promise, then a withdrawal unsettled earnings amount will be
overriten incorrectly. We now simply ignore the withdrawal promise.