Increase withdrawal amount to 500 MYST
We got a little over 100 consumer with balances above 100 MYST.
To make withdrawals for them easier we can increase the total withdrawal
amount to 500 MYST. This fixes our problem with requiring us to do
multiple withdrawal for certain consumers. Above this amount there seems
to be only consumers who commited fraud, so we do not care about that.