Pass transactor errors to the caller if possible
If possible we should pass transactor errors to the caller.
This allows us to return custom errors straight from transactor if
needed. For example we should be able to handle fee errors with custom
`code` field.

This change comes before actual new values are being returned from
transactor as to no break the UI too much.